Refactoring, Borradp de la carpeta Catalog y pasado todo su contenido a la carpeta Skills
