+++ src/app/app.module.ts
@@ -15,7 +15,7 @@ import { MessageService } from 'primeng/api';
 import { DialogService } from 'primeng/dynamicdialog';
 import { HttpInterceptorService } from './core/services/http-interceptor.service';
 import { RefreshTokenResolverService } from './core/services/refresh-token-resolver.service';
-import { CatalogModule } from './catalog/catalog.module';
+
 import { TrainingModule } from './training/training.module';
 
 registerLocaleData(localeEs, 'es');
@@ -31,7 +31,6 @@ registerLocaleData(localeEs, 'es');
     LoginModule,
     CardModule,
     SkillsModule,
-    CatalogModule,
     TrainingModule,
   ],
   providers: [
