+++ src/app/skills/capabilities/capabilities-list/capabilities-list.component.scss
