+++ src/app/skills/capabilities/capabilities-upload/capabilities-upload.component.html
