+++ src/app/skills/capabilities/model/Capability.ts
