+++ src/app/skills/certifications/certifications-list/certifications-list.component.scss
