+++ src/app/skills/report/model/Report.ts
