+++ src/app/skills/report/report-list/report-list.component.scss
